separated controller from RunningGameState
