+++ modules/states/play.py
@@ -8,6 +8,7 @@ from modules.ai.ai import Ai
 from modules.ai.vectors import VectorCalculator
 from modules.behavior.camera import Camera
 from modules.behavior.collisions import Collisions
+from modules.behavior.controller import Controller
 from modules.objects.player import Player
 from modules.objects.tiles import Ground
 from modules.settings import CATCHERS, RUNNERS, SCREEN_HEIGHT, SCREEN_WIDTH, TILE_SIZE
@@ -119,49 +120,3 @@ class RunningGameState(GameState):
             squares.append(square)
 
         return squares
-
-
-class Controller:
-    """Obsługa sterowania w grze."""
-
-    def __init__(self, squares, game_state, camera):
-        """Inicjalizuje kontroler z danymi kwadratami."""
-        self.squares = squares
-        self.game_state = game_state
-        self.camera = camera
-
-    def handle_movement(self):
-        """Obsługuje zdarzenia związane z ciągłym naciśnięciem klawisza."""
-        keys = pygame.key.get_pressed()
-
-        key_handlers = {
-            pygame.K_a: self.squares.move_left,
-            pygame.K_LEFT: self.squares.move_left,
-            pygame.K_d: self.squares.move_right,
-            pygame.K_RIGHT: self.squares.move_right,
-        }
-
-        for key, handler in key_handlers.items():
-            if keys[key]:
-                handler()
-
-    def handle_key_release(self, event):
-        """Obsługuje zdarzenia związane z puszczeniem klawisza."""
-        if event.key in (pygame.K_a, pygame.K_LEFT, pygame.K_d, pygame.K_RIGHT):
-            self.squares.velocity_x = 0  # Zresetuj prędkość x kwadratu
-
-    def handle_key_press_actions(self, event):
-        """Obsługuje zdarzenia związane z naciśnięciem klawisza."""
-        if event.key == pygame.K_ESCAPE:
-            return self.game_state.pause_menu_state
-        elif event.key == pygame.K_SPACE:
-            self.squares.jump()
-
-    def handle_quit_event(self, event):
-        """Obsługuje zdarzenie wyjścia z gry."""
-        pygame.quit()
-        quit()
-
-    def set_pause_state(self, pause_state):
-        """Ustawia stan pauzy dla stanu gry."""
-        self.pause_state = pause_state
